refactor(std): move 7 platform-divergent globals to rts-std
console performance headers form_data event_target message_channel abort -> rts-std/src/globals/. São backend (console->io, performance->time, abort->error slot; headers/form_data/event_target/message_channel modelam recursos de plataforma).

Refs reescritas: gc::handles -> rts_engine::heap::handles. abort chama gc::error::__RTS_FN_RT_ERROR_SET via extern-decl local (no_mangle no collector, resolve por link). event_target/message_channel/abort já declaravam __RTS_FN_RT_INVOKE_AUTO como extern (no_mangle em rts-shared function/ops) — resolve por link sem dep de rts-shared.

Cargo rts-std += indexmap. Facade no runtime globals/mod.rs via pub use rts_std::globals::*. Codegen inalterado.

DEFERIDOS (bloqueados por pub-fn não-extern-able): timers (text_encoding::drain_microtasks), fetch (text_encoding microtask + promise_slot + ureq), blob/readable_stream (promise_slot::new_fulfilled + flate2). Movem com text_encoding/promise_slot (steps 5/6).

Gate: build limpo; rts.exe test 1710/1710; AOT smoke (console/performance/Headers/FormData/EventTarget/AbortController, JIT==AOT) OK. NB pré-existente: abort() sem arg tem arity strict (spec pede reason); abort("x") funciona.
