+++ crates/rts-runtime/src/namespaces/globals/mod.rs
@@ -2,15 +2,15 @@
 //! + JSON, Date, console, globalThis, RegExp, Error family, timers, fetch,
 //! TextEncoder/Decoder, atob/btoa, structuredClone, URL, performance.
 
-pub mod abort;
+pub use rts_std::globals::abort;
 pub use rts_shared::globals::bigint;
 pub mod blob;
 pub mod dataview;
 pub use rts_shared::globals::dom_exception;
-pub mod event_target;
-pub mod form_data;
+pub use rts_std::globals::event_target;
+pub use rts_std::globals::form_data;
 pub use rts_shared::globals::boolean;
-pub mod console;
+pub use rts_std::globals::console;
 pub use rts_shared::globals::date;
 pub use rts_shared::globals::number;
 pub use rts_shared::globals::error;
@@ -18,12 +18,12 @@ pub mod events;
 pub mod fetch;
 pub use rts_shared::globals::function;
 pub use rts_shared::globals::global_this;
-pub mod headers;
+pub use rts_std::globals::headers;
 pub use rts_shared::globals::intl;
 pub use rts_shared::globals::json;
 pub use rts_shared::globals::json5;
-pub mod message_channel;
-pub mod performance;
+pub use rts_std::globals::message_channel;
+pub use rts_std::globals::performance;
 pub use rts_shared::globals::proxy;
 pub mod readable_stream;
 pub use rts_shared::globals::reflect;
