+++ crates/rts-std/Cargo.toml
@@ -13,6 +13,7 @@ asio = ["cpal/asio"]
 
 [dependencies]
 rts-engine = { path = "../rts-engine" }
+indexmap = "2.14.0"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync"] }
 actix-web = { version = "4", default-features = false, features = ["macros"] }
 actix-rt = "2"
