+++ crates/rts-std/src/globals/console/mod.rs
