+++ crates/rts-std/src/globals/console/rt.rs
