+++ crates/rts-std/src/globals/event_target/mod.rs
@@ -14,7 +14,7 @@ use indexmap::IndexMap;
 use rts_engine::abi::ty::{Bool, Handle};
 use rts_engine::{AbiType, Engine, FnPtr, Member, MemberFlags, MemberKind, Sig};
 
-use crate::namespaces::gc::handles::{alloc_entry, with_entry, with_entry_mut, Entry};
+use rts_engine::heap::handles::{alloc_entry, with_entry, with_entry_mut, Entry};
 
 // ── EventTarget ───────────────────────────────────────────────────────────────
