+++ crates/rts-std/src/globals/form_data/mod.rs
@@ -11,7 +11,7 @@ use indexmap::IndexMap;
 use rts_engine::abi::ty::{Bool, Handle};
 use rts_engine::{AbiType, Engine, FnPtr, Member, MemberFlags, MemberKind, Sig};
 
-use crate::namespaces::gc::handles::{alloc_entry, with_entry, with_entry_mut, Entry};
+use rts_engine::heap::handles::{alloc_entry, with_entry, with_entry_mut, Entry};
 
 /// Storage interno: handle do Vec<i64> de pares (key_h, val_h).
 fn pairs_handle(h: u64) -> u64 {
