+++ crates/rts-std/src/globals/message_channel/mod.rs
@@ -16,7 +16,7 @@ use indexmap::IndexMap;
 use rts_engine::abi::ty::Handle;
 use rts_engine::{AbiType, Engine, FnPtr, Member, MemberFlags, MemberKind, Sig};
 
-use crate::namespaces::gc::handles::{alloc_entry, with_entry, with_entry_mut, Entry};
+use rts_engine::heap::handles::{alloc_entry, with_entry, with_entry_mut, Entry};
 
 fn map_get(h: u64, key: &str) -> i64 {
     with_entry(h, |e| match e {
