+++ crates/rts-std/src/globals/performance/mod.rs
