+++ crates/rts-std/src/lib.rs
@@ -19,3 +19,4 @@ pub mod fs;
 pub mod tls;
 pub mod thread;
 pub mod http_server;
+pub mod globals;
